+++ app/divisions/[slug]/page.tsx
@@ -1,10 +1,32 @@
 import React from "react";
+import PageLayout from "app/components/pageLayout/pageLayout";
+import { Metadata } from "next";
 
-export default function Division() {
+type DivisionProps = {
+  params: { slug: string };
+};
+
+export async function generateMetadata({
+  params,
+}: DivisionProps): Promise<Metadata> {
+  const slug = params.slug;
+  return {
+    title: `${slug} - NYPL Digital Collections`,
+  };
+}
+
+export default function Division({ params }: DivisionProps) {
   return (
-    <>
-      <title>[Name of Division] - NYPL Digital Collections</title>
-      <h2> Division name </h2>
-    </>
+    <PageLayout
+      activePage="division"
+      breadcrumbs={[
+        { text: "Home", url: "/" },
+        { text: "All Divisions", url: "/divisions" },
+        { text: `${params.slug} Division`, url: `/divisions/${params.slug}` },
+      ]}
+    >
+      <h2> {params.slug} Division </h2>
+      <br />
+    </PageLayout>
   );
 }